Rename `plugins` to `plugin`
