+++ client-js/build.gradle.kts
@@ -29,9 +29,9 @@ import com.google.protobuf.gradle.testProtobuf
 import io.spine.internal.gradle.base.check
 import io.spine.internal.gradle.fs.LazyTempPath
 import io.spine.internal.gradle.javascript.javascript
-import io.spine.internal.gradle.javascript.plugins.idea
-import io.spine.internal.gradle.javascript.plugins.mcJs
-import io.spine.internal.gradle.javascript.plugins.protobuf
+import io.spine.internal.gradle.javascript.plugin.idea
+import io.spine.internal.gradle.javascript.plugin.mcJs
+import io.spine.internal.gradle.javascript.plugin.protobuf
 import io.spine.internal.gradle.javascript.task.assemble
 import io.spine.internal.gradle.javascript.task.check
 import io.spine.internal.gradle.javascript.task.clean
